scraper/requirements.txt
playwright==1.44.0
APScheduler==3.10.4
supabase==2.5.0
httpx==0.27.0
selectolax==0.3.21
python-dotenv==1.0.1
